requirements.txt
casttube==0.2.0
certifi==2018.11.29
chardet==3.0.4
colorgram.py==1.1.0
cycler==0.10.0
idna==2.8
ifaddr==0.1.6
Pillow==5.3.0
protobuf==3.6.1
psutil==5.4.8
PyChromecast==2.3.0
pyparsing==2.3.1
python-dateutil==2.7.5
requests==2.21.0
six==1.12.0
urllib3==1.24.1
zeroconf==0.21.3
